# Conflicts:
#	core/src/main/java/de/jplag/SubsequenceHashLookupTable.java
#	languages.api/src/main/java/de/jplag/TokenHashMap.java
#	languages.api/src/main/java/de/jplag/TokenList.java
#	languages/text/pom.xml
#	languages/text/src/main/antlr/text.g
#	languages/text/src/main/java/de/jplag/text/AntlrParserToken.java
#	languages/text/src/main/java/de/jplag/text/InputState.java

+++ core/src/main/java/de/jplag/reporting/reportobject/model/Cluster.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record Cluster(@JsonProperty("average_similarity") float averageSimilarity, @JsonProperty("strength") float strength,
+public record Cluster(@JsonProperty("average_similarity") double averageSimilarity, @JsonProperty("strength") double strength,
         @JsonProperty("members") List<String> members) {
 }
